event-sources: Refactor object pkg (#6511)
* Use common functional option for all object types

Removes redundancy by allowing the definition of options that work for
all objects (e.g. methods that alter the objects' metadata).

* Remove the term "Container" from Service options

It can be inferred from the context that images, env vars and ports are
set on containers.

* Remove object builders from testing pkg

Replace with explicit object struct definitions. "testing" builders
overlap with the builders from the "object" package and are very
opinionated, making them unlikely to be reusable in future tests.