Event Service Compatibility Layer (#7147)
Related issues
#7091.

Event Service Compatibility Layer redirects the request to the new Knative Eventing Mesh. this respects the public contract exposed previously to all the external Applications.
With this change the request received to the event service will not be propagated to the Event-bus publish service. Hence Modified the relevant tests as part of this PR.

Application Operator Change: #7150
Image Bump: #7153