Retry reconciling subscriptions on NATS disconnect
If the NATS server is down and the NATS subscription reconciler exceeds the retry
attempts configured, the Kyma subscription would still be reported as ready, even
though no messages could be delivered. The PR fixes this by forcing reconciliation
of the Kyma subscriptions after reconnection attempts have been exhausted. This
would move the subscriptions to a not ready state, which would cause periodic reconciliation
of the subscriptions until NATS is available again.