Fix fast-integration tests on upgrade jobs (#14282)
* Fix fast-integration tests on upgrade jobs

PR #14103 moved flags from the Grafana auth proxy deployment's cli flags
to a secret to make it better configurable in tests. This change led to
failing upgrade tests since adding the variables to the secret was not
effective on the old version of the deployment.

This PR partially reverts the way how the configuration is applyed in
the integration tests to make it working with both auth proxy
deployments.

* add import

* update resetGrafanaProxy

* add isSkr env

* remove --reverse-proxy=true

* remove if skr in eventing test

* change to isSkr variable

Co-authored-by: Dennis Geiselmann <dennis.geiselmann@sap.com>